refactor: rename tests to follow camelCase, seperate them into services and components and fix imports
