+++ frontend/tests/components/legend.test.tsx
@@ -1,7 +1,7 @@
 import "@testing-library/jest-dom";
 import { render, screen } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
-import Legend from "../src/components/Legend";
+import Legend from "../../src/components/Legend";
 
 describe("Legend", () => {
   it("renders without crashing", () => {
