+++ frontend/tests/components/personForm.test.tsx
@@ -1,16 +1,16 @@
 import "@testing-library/jest-dom";
 import { fireEvent, render, screen, waitFor } from "@testing-library/react";
 import { afterEach, describe, expect, it, vi } from "vitest";
-import PersonForm from "../src/components/PersonForm.tsx";
+import PersonForm from "../../src/components/PersonForm.tsx";
 
-vi.mock("../src/services/peopleService", () => ({
+vi.mock("../../src/services/peopleService", () => ({
   findAllPeople: vi.fn().mockResolvedValue([
     { id: 1, firstName: "Joku", lastName: "Esihenkilö" },
     { id: 2, firstName: "Muu", lastName: "Esihenkilö" },
   ]),
 }));
 
-vi.mock("../src/services/referenceDataService", () => ({
+vi.mock("../../src/services/referenceDataService", () => ({
   findAllDepartments: vi.fn().mockResolvedValue([
     { id: 1, name: "IT" },
     { id: 2, name: "HR" },
