+++ frontend/tests/components/settingsModal.test.tsx
@@ -2,7 +2,7 @@ import "@testing-library/jest-dom";
 import { fireEvent, render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import SettingsModal from "../src/components/SettingsModal.tsx";
+import SettingsModal from "../../src/components/SettingsModal.tsx";
 
 describe("SettingsModal", () => {
   const onCloseMock = vi.fn();
