+++ frontend/tests/components/topBar.test.tsx
@@ -2,16 +2,16 @@ import "@testing-library/jest-dom";
 import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { describe, expect, it, vi } from "vitest";
-import TopBar from "../src/components/TopBar.tsx";
+import TopBar from "../../src/components/TopBar.tsx";
 
-vi.mock("../src/hooks/useAuth", () => ({
+vi.mock("../../src/hooks/useAuth", () => ({
   useAuth: () => ({
     user: { name: "Test User" },
     logout: vi.fn(),
   }),
 }));
 
-vi.mock("../src/hooks/useRoomSelection", () => ({
+vi.mock("../../src/hooks/useRoomSelection", () => ({
   useRoomSelection: () => ({
     activeRoomId: null,
     setActiveRoomId: vi.fn(),
